+++ cli/commands/usage.go
@@ -66,6 +66,8 @@ script for the RPM and DEB packages.`
 
 	postInstUsage = `Path to the file that contains post install
 script for the RPM and DEB packages.`
+
+	UnitParamsFileUsage = `Path to the file that contains systemd unit params`
 )
 
 // RUNNING